requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
mailslurp-client==15.19.22
MarkupSafe==2.1.5
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
